% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_setdiff}
\alias{arma_setdiff}
\title{Calculating differences between two sets}
\usage{
arma_setdiff(x, y)
}
\arguments{
\item{x}{vector of elements in first set}

\item{y}{vector of elements in second set}
}
\value{
Vector of set differences
}
\description{
Calculating differences between two sets
}
\examples{
x <- c(1,3,4,5,8); y <- c(1, 2, 4, 5, 8)
arma_setdiff(x,y)
}
