% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ArmaSample}
\alias{ArmaSample}
\title{Sampling with replacement from a vector of numbers}
\usage{
ArmaSample(x, numObs, weights)
}
\arguments{
\item{x}{Vector of elements to sample from}

\item{numObs}{Number of draws required}

\item{weights}{Sampling weights}
}
\value{
vector of draws
}
\description{
Sampling with replacement from a vector of numbers
}
