% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVselectC}
\alias{CVselectC}
\title{Fits the VNPDA model for a set of candidate value of smoothing parameters for each variable and returns the optimal smoothing parameter value}
\usage{
CVselectC(xtrain, ytrain, cval, ay, by, ag, bg, maxdepth = 10L)
}
\arguments{
\item{xtrain}{Variables of training sample (matrix)}

\item{ytrain}{Group labels of training sample}

\item{cval}{candidate values for smoothing parameter (matrix)}

\item{ay}{Hyparameter 1 of Beta prior for rho_y}

\item{by}{Hyparameter 2 of Beta prior for rho_y}

\item{ag}{Hyparameter 1 of Beta prior for rho_gamma}

\item{bg}{Hyparameter 2 of Beta prior for rho_gamma}

\item{maxdepth}{Maximum depth to truncate the tree. Default is 10.}

\item{Xtest}{Variables of testing sample for classification (matrix)}
}
\value{
Optimal smoothing parameter settings
}
\description{
Fits the VNPDA model for a set of candidate value of smoothing parameters for each variable and returns the optimal smoothing parameter value
}
