% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DividePredCpp}
\alias{DividePredCpp}
\title{Calculate classification probabilities for binary Polya tree classification problem with a single feature}
\usage{
DividePredCpp(x1, x0, xnew, cval, m, s, DOWN = 0, UP = 1, depth = 0L,
  maxdepth = 10L)
}
\arguments{
\item{x1}{training sample from group 1}

\item{x0}{training sample from group 0}

\item{xnew}{new data point}

\item{cval}{smoothing parameter}

\item{m}{mean of centering Gaussian distribution}

\item{s}{sd of centering Gaussian distribution}

\item{DOWN}{Leave this number as 0}

\item{UP}{Leave this number as 1}

\item{depth}{Leave this number as 0}

\item{maxdepth}{Maximum depth to truncate the tree. Default is 10.}
}
\value{
Classification probability
}
\description{
Calculate classification probabilities for binary Polya tree classification problem with a single feature
}
\examples{
samp1 <- rnorm(50,4,1)
samp0 <- rt(40,3)
xtest <- rt(1,3)
DividePredCpp(x1 = samp1, x0 = samp0, xnew = xtest, cval = 1, m = mean(c(samp1,samp0)), s = sd(c(samp1,samp0)))
}
