% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epsilonC}
\alias{epsilonC}
\title{Applies the epsilon function in Yu et. al. (2018).}
\usage{
epsilonC(v)
}
\arguments{
\item{v}{vector of input}
}
\value{
Vector of output
}
\description{
Applies the epsilon function in Yu et. al. (2018).
}
\examples{
x <- rt(1,10)
epsilonC(x)
}
