% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expit}
\alias{expit}
\title{Applies the expit (inverse logit) function.}
\usage{
expit(v)
}
\arguments{
\item{v}{vector of input}
}
\value{
Vector of output
}
\description{
Applies the expit (inverse logit) function.
}
\examples{
x <- rt(100,1)
expit(x)
}
