% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTmethods6test.R
\name{Ginv}
\alias{Ginv}
\title{G inverse calculation

Describe}
\usage{
Ginv(p, xbar, s)
}
\arguments{
\item{p}{Numeric vector of probabilities}

\item{xbar}{Numeric Gaussian mean}

\item{s}{Numeric Gaussian sd}
}
\value{
Numeric vector
}
\description{
G inverse calculation

Describe
}
\examples{
q_vec <- c(1:5)/6
Ginv(q_vec,xbar=0,s=1)

}
