% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTmethods6test.R
\name{HistoTree3}
\alias{HistoTree3}
\title{Calculate recursive histograms}
\usage{
HistoTree3(x1, x0 = NULL, prev_n1 = 0, prev_n0 = 0, c, DOWN, UP, m1,
  s1, m0 = NULL, s0 = NULL, depth = 0, maxdepth = 20,
  samples = 1)
}
\arguments{
\item{x1}{Numeric vector for sample 1}

\item{x0}{Numeric vector for sample 2}

\item{prev_n1}{Numeric}

\item{prev_n0}{Numeric}

\item{c}{Numeric}

\item{DOWN}{Numeric}

\item{UP}{Numeric}

\item{m1}{Numeric}

\item{s1}{Numeric}

\item{m0}{Numeric}

\item{s0}{Numeric}

\item{depth}{Numeric}

\item{maxdepth}{Numeric}

\item{samples}{Numeric for number of samples}
}
\value{
a list object
}
\description{
Calculate recursive histograms
}
\examples{
test_samp <- rnorm(60,0,3)
HistoTree3(x1=test_samp,DOWN=0,c=3,UP=1,m1=mean(test_samp),s1=sd(test_samp))
}
