% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSmatrix}
\alias{KSmatrix}
\title{Calls the two-sample KS test function for two multivariate samples in R}
\usage{
KSmatrix(X, y)
}
\arguments{
\item{X}{Sample}

\item{y}{Group labels (1 or 0)}
}
\value{
KS test p-value
}
\description{
Calls the two-sample KS test function for two multivariate samples in R
}
