% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PerturbCpp}
\alias{PerturbCpp}
\title{Introduces random perturbation to a set of variables}
\usage{
PerturbCpp(Z)
}
\arguments{
\item{Z}{matrix of values to be perturbed}
}
\value{
perturbed matrix
}
\description{
Introduces random perturbation to a set of variables
}
