% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTmethods6test.R
\name{PlotPT}
\alias{PlotPT}
\title{Plots Polya Tree Predictive Density}
\usage{
PlotPT(x1, x0 = NULL, c, maxdepth = 10, mainTitle = NULL,
  HorizLab = NULL, x_limits = NULL)
}
\arguments{
\item{x1}{Numeric vector sample 1}

\item{x0}{Numeric vector sample 1}

\item{c}{Numeric smoothing parameter}

\item{maxdepth}{Numeric maximum tree depth}

\item{mainTitle}{Text for plot heading}

\item{HorizLab}{Text for x-axis label}

\item{x_limits}{Limits of x-axis}
}
\value{
ggplot object
}
\description{
Plots Polya Tree Predictive Density
}
\examples{
x_samp <- rnorm(30,3,2)
y_samp <- rt(30,2)
PlotPT(x1=x_samp, x0 = y_samp, c=2)
}
