% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PredCpp}
\alias{PredCpp}
\title{Calculate classification probabilities for binary Polya tree classification problem with multiple features}
\usage{
PredCpp(xtrain1, xtrain0, xtest, cval, omega, ay, by, maxdepth = 10L)
}
\arguments{
\item{xtrain1}{training sample from group 1 (matrix)}

\item{xtrain0}{training sample from group 0 (matrix)}

\item{xtest}{a single new data point (vector)}

\item{cval}{smoothing parameters}

\item{omega}{vector of variable selection weights}

\item{ay}{hyperparameter 1 of beta prior}

\item{by}{hyperparameter 2 of beta prior}

\item{maxdepth}{Maximum depth to truncate the tree. Default is 10.}
}
\value{
Classification probability
}
\description{
Calculate classification probabilities for binary Polya tree classification problem with multiple features
}
