% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTmethods6test.R
\name{PT_Pred_Density}
\alias{PT_Pred_Density}
\title{Calculates Polya Tree Predictive Density}
\usage{
PT_Pred_Density(x, probs, cutOFFs, m = 0, s = 1)
}
\arguments{
\item{x}{Numeric vector of sample to be plotted}

\item{probs}{Numeric vector of probabilities}

\item{cutOFFs}{Bin cutoffs}

\item{m}{Numeric mean of centering distribution}

\item{s}{Numeric sd of centering distribution}
}
\value{
Numeric vector of predictive density
}
\description{
Calculates Polya Tree Predictive Density
}
