% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleMixNorm}
\alias{SampleMixNorm}
\title{Sampling from a mixture of univariate normals}
\usage{
SampleMixNorm(N, p, indices, weights, meanVec, sdVec)
}
\arguments{
\item{N}{Number of rows in output}

\item{p}{Number of columns in outout}

\item{indices}{Recommended to be sequence from 1 to p}

\item{weights}{Sampling probabilities}

\item{meanVec}{Vector of Gaussian means}

\item{sdVec}{Vector of Gaussian SDs}
}
\value{
matrix N by p of draws
}
\description{
Sampling from a mixture of univariate normals
}
