% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simpleKS}
\alias{simpleKS}
\title{Calls the two-sample KS test function for two univariate samples in R}
\usage{
simpleKS(x, y)
}
\arguments{
\item{x}{Sample 1}

\item{y}{Sample 0}
}
\value{
KS test p-value
}
\description{
Calls the two-sample KS test function for two univariate samples in R
}
