% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TREECppCVselect}
\alias{TREECppCVselect}
\title{Performs a two-sample Bayesian nonparametric Polya tree test between two independent univariate samples (for multiple candidate values of smoothing paramater)}
\usage{
TREECppCVselect(x1, x0, cval, m, s, DOWN = 0, UP = 1, depth = 0,
  maxdepth = 10L)
}
\arguments{
\item{x1}{training sample from group 1}

\item{x0}{training sample from group 0}

\item{cval}{candidate value for smoothing parameter (vector)}

\item{m}{mean of centering Gaussian distribution}

\item{s}{sd of centering Gaussian distribution}

\item{DOWN}{Leave this number as 0}

\item{UP}{Leave this number as 1}

\item{depth}{Leave this number as 0}

\item{maxdepth}{Maximum depth to truncate the tree. Default is 10.}
}
\value{
log Bayes factor
}
\description{
Performs a two-sample Bayesian nonparametric Polya tree test between two independent univariate samples (for multiple candidate values of smoothing paramater)
}
