% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VQDA}
\alias{VQDA}
\title{Fit the VQDA model in Yu et. al. (2018).}
\usage{
VQDA(vy, mX, mXtest, r, kappa)
}
\arguments{
\item{vy}{Class labels of training sample}

\item{mX}{Variables of training sample}

\item{mXtest}{Variables of testing sample (for prediction)}

\item{r}{tuning parameter 1 for penalty  on log likelihood}

\item{kappa}{tuning parameter 2 for penalty  on log likelihood}
}
\value{
List containing variable selection probabilities, classification probabilitiew, likelihood ratio statistics and classification results
}
\description{
Fit the VQDA model in Yu et. al. (2018).
}
\examples{
data(LIHCtoy)
X <- as.matrix(X)
resVQDA <- VQDA(vy = y[1:90], mX = X[1:90,], mXtest = X[91:109,], r = 0.98, kappa = (1/7))
resVQDA
}
